Revert "Update modele"
This reverts commit f42bcf05b40be71172abd7f89d297478b7dad52b.

+++ src/tsp/Graph.java
@@ -9,12 +9,12 @@ public interface Graph {
 	/**
 	 * @return the maximal cost of an arc of <code>this</code>
 	 */
-	public abstract int getMaxArcCost();
+	public abstract double getMaxArcCost();
 
 	/**
 	 * @return the minimal cost of an arc of <code>this</code>
 	 */
-	public abstract int getMinArcCost();
+	public abstract double getMinArcCost();
 
 	/**
 	 * @return the number of vertices of <code>this</code>
@@ -26,7 +26,7 @@ public interface Graph {
 	 * if <code>(i,j)</code> is an arc of <code>this</code>, then <code>cost[i][j]</code> = cost of <code>(i,j)</code>, 
 	 * otherwise <code>cost[i][j] = this.getMaxArcCost()+1</code>
 	 */
-	public abstract int[][] getCost();
+	public abstract double[][] getCost();
 
 	/**
 	 * @param i a vertex such that <code>0 <= i < this.getNbVertices()</code>
